+++ ui/src/App/App.tsx
@@ -1,5 +1,5 @@
 import React, {Component} from 'react'
-import NavBar from "./NavBar"
+import NavBar from "../NavBar"
 
 class App extends Component {
     render() {