+++ ui/src/App/Dependencies/Dependencies.ts
@@ -1,0 +1,3 @@
+export default interface Dependencies {
+    ws: SocketIOClient.Socket
+}