+++ ui/src/App/Dependencies/getDefaultDependencies.ts
@@ -1,0 +1,10 @@
+import openSocket from 'socket.io-client'
+import Dependencies from "./Dependencies"
+
+export default function getDefaultDependencies(): Dependencies {
+    const socket = openSocket('http://localhost:8777')
+
+    return {
+        ws: socket,
+    }
+}