+++ ui/src/index.tsx
@@ -1,14 +1,37 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import React from 'react'
+import ReactDOM from 'react-dom'
+import { BrowserRouter } from 'react-router-dom'
+import {Provider} from 'react-redux'
+import './index.css'
+import App from './App/App'
+import * as serviceWorker from './serviceWorker'
 
 import 'normalize.css/normalize.css'
 import '@blueprintjs/icons/lib/css/blueprint-icons.css'
 import '@blueprintjs/core/lib/css/blueprint.css'
+import getDefaultDependencies from "./App/Dependencies/getDefaultDependencies"
+import makeStore from "./App/Store/makeStore"
+import registerSubscribers from "./App/Subscriptions/registerSubscribers"
 
-ReactDOM.render(<App />, document.getElementById('root'));
+// import openSocket from 'socket.io-client'
+// const socket = openSocket('http://localhost:8777')
+//
+// function subscribeToTimer(cb) {
+//     socket.on('timer', timestamp => cb(null, timestamp))
+//     socket.emit('subscribeToTimer', 1000)
+// }
+const deps = getDefaultDependencies()
+const store = makeStore(deps)
+
+// Register WS subscribers
+registerSubscribers(store, deps)
+
+ReactDOM.render(
+    <Provider store={store}>
+        <BrowserRouter>
+            <App />
+        </BrowserRouter>
+    </Provider>, document.getElementById('root'))
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.