+++ ui/src/App/PageNotFound.tsx
@@ -1,0 +1,9 @@
+import React, { PureComponent } from 'react'
+
+class PageNotFound extends PureComponent {
+    render() {
+        return <div>Not Found</div>
+    }
+}
+
+export default PageNotFound