+++ ui/src/Dashboard/Dashboard.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react'
+
+const Dashboard = () => (
+    <div>
+        Dashboard
+    </div>
+)
+
+export default Dashboard
+