+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!--
 ## [Unreleased]
+-->
+
+## [0.4.0] - 2019-02-06
 ### Added
 - UI Scaffolding
 - Implementing logging to file (`~/.jetpax/logs/server/server.log`) for Jetpax internals
@@ -78,7 +82,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial repository contents
 
-[Unreleased]: https://github.com/andyfleming/jetpax/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/andyfleming/jetpax/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/andyfleming/jetpax/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/andyfleming/jetpax/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/andyfleming/jetpax/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/andyfleming/jetpax/compare/v0.1.1...v0.2.0